removed all of the tag pages and dynamic pages created from it. added shop page and changed links to point to shop
