+++ src/pages/shop.js
@@ -1,0 +1,67 @@
+import styled from '@emotion/styled';
+import { graphql, Link } from 'gatsby';
+import PropTypes from 'prop-types';
+import React from 'react';
+import Layout from '../components/layout';
+import TextLink from '../components/links/text-link';
+import SEO from '../components/seo';
+import { blogMenuLinks } from '../components/_config/menu-links';
+import { StyledH1 } from '../components/_shared/styled-headings';
+import { StyledFullHeightSection } from '../components/_shared/styled-section';
+import { StyledSeparator } from '../components/_shared/styled-separator';
+
+const StyledTagsH1 = styled(StyledH1)`
+  margin-top: 3rem;
+`;
+
+const Shop = ({
+  data: {
+    allMarkdownRemark: { group },
+  },
+}) => {
+
+  return (
+    <Layout menuLinks={blogMenuLinks}>
+      <SEO title="Shop" />
+      <StyledFullHeightSection>
+        <StyledTagsH1>Shop</StyledTagsH1>
+        <StyledSeparator />
+        This is where you can buy stuff.
+        <StyledSeparator />
+        <TextLink label="Take me home" link="/" />
+      </StyledFullHeightSection>
+    </Layout>
+  );
+};
+
+Shop.propTypes = {
+  data: PropTypes.shape({
+    allMarkdownRemark: PropTypes.shape({
+      group: PropTypes.arrayOf(
+        PropTypes.shape({
+          fieldValue: PropTypes.string.isRequired,
+          totalCount: PropTypes.number.isRequired,
+        }).isRequired
+      ),
+    }),
+    site: PropTypes.shape({
+      siteMetadata: PropTypes.shape({
+        title: PropTypes.string.isRequired,
+      }),
+    }),
+  }),
+};
+export default Shop;
+
+export const pageQuery = graphql`
+  query {
+    allMarkdownRemark(
+      filter: { fileAbsolutePath: { regex: "/content/posts/" }, frontmatter: { published: { eq: true } } }
+    ) {
+      group(field: frontmatter___tags) {
+        fieldValue
+        totalCount
+      }
+    }
+  }
+`;
