+++ src/templates/blog-post.js
@@ -4,7 +4,6 @@ import Img from 'gatsby-image';
 import PropTypes from 'prop-types';
 import React from 'react';
 import Layout from '../components/layout';
-import TagList from '../components/tag-list';
 import { blogMenuLinks } from '../components/_config/menu-links';
 import { StyledH1 } from '../components/_shared/styled-headings';
 import { StyledSection } from '../components/_shared/styled-section';
@@ -39,7 +38,7 @@ const BlogPost = ({ data }) => {
   const readingTime = data.markdownRemark.fields.readingTime.text;
   const post = data.markdownRemark;
   const coverImage = post.frontmatter.cover_image ? post.frontmatter.cover_image.childImageSharp.fluid : null;
-  const { tags = [], title, date } = post.frontmatter;
+  const { title, date } = post.frontmatter;
 
   return (
     <Layout menuLinks={blogMenuLinks}>
@@ -48,7 +47,6 @@ const BlogPost = ({ data }) => {
         <StyledDate>
           Posted {date}. <span>{readingTime}.</span>
         </StyledDate>
-        <TagList tags={tags} />
         {coverImage && <Img fluid={coverImage} />}
         <StyledBlogText dangerouslySetInnerHTML={{ __html: post.html }} />
       </StyledBlogSection>
